Gold coins are now tracked by bot. It has an inventory array that stores the items, a bot.bucks that counts the value, a bot.inventoryCap that counts the number and decrements on picking it up and an alert fires and the pickup fails if you are over capacity.
NEXT STEPS!!!
1. Figured out how to implement for loops! The bCrumb.persistent shouldn't be
a boolean, it should be a counter that counts down interations until it
disapears.

2. After this, consider a mechanic that allows you to stack bCrumbs and
have only the top one be active. That way "for" loops can run a function
once they are complete.

3. If/else -- maybe (if inventory === 0) to contorl entry/ exit into a
loop. Else is not needed, just use two "if"s stacked after each other.
This would need a direction to be combined with the operator.
	modified:   classes.js
	modified:   easel-source.js